Matched varibale types 1to1 with SQL

+++ pokepi/src/models/ability.rs
@@ -2,10 +2,10 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Ability {
-    pub ability_id: u16,
+    pub ability_id: i32,
     pub ability_name: String,
     pub ability_identifier: String,
-    pub generation_introduced: u8,
+    pub generation_introduced: i16,
     pub is_main_series: bool,
 }
 
@@ -13,7 +13,7 @@ pub struct Ability {
 pub struct CreateAbility {
     pub ability_name: String,
     pub ability_identifier: String,
-    pub generation_introduced: u8,
+    pub generation_introduced: i16,
     pub is_main_series: bool,
 }
 
@@ -21,6 +21,6 @@ pub struct CreateAbility {
 pub struct UpdateAbility {
     pub ability_name: Option<String>,
     pub ability_identifier: Option<String>,
-    pub generation_introduced: Option<u8>,
+    pub generation_introduced: Option<i16>,
     pub is_main_series: Option<bool>,
 }
