+++ pokepi/src/models/egg_group.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct EggGroup {
-    pub egg_group_id: u8,
+    pub egg_group_id: i32,
     pub egg_group_name: String,
     pub egg_group_identifier: String,
 }
