+++ pokepi/src/models/encounter_condition.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct EncounterCondition {
-    pub encounter_id: u32,
-    pub encounter_condition_value_id: u16,
+    pub encounter_id: i32,
+    pub encounter_condition_value_id: i32,
 }
 
 pub type CreateEncounterCondition = EncounterCondition;
