+++ pokepi/src/models/game.rs
@@ -2,13 +2,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Game {
-    pub game_id: u16,
+    pub game_id: i32,
     pub game_name: String,
     pub game_identifier: String,
-    pub generation: u8,
-    pub version_group_id: u16,
+    pub generation: i16,
+    pub version_group_id: i32,
     pub release_date: String,
-    pub platform_id: u16,
+    pub platform_id: i32,
     pub is_main_series: bool,
 }
 
@@ -16,10 +16,10 @@ pub struct Game {
 pub struct CreateGame {
     pub game_name: String,
     pub game_identifier: String,
-    pub generation: u8,
-    pub version_group_id: u16,
+    pub generation: i16,
+    pub version_group_id: i32,
     pub release_date: String,
-    pub platform_id: u16,
+    pub platform_id: i32,
     pub is_main_series: bool,
 }
 
@@ -27,9 +27,9 @@ pub struct CreateGame {
 pub struct UpdateGame {
     pub game_name: Option<String>,
     pub game_identifier: Option<String>,
-    pub generation: Option<u8>,
-    pub version_group_id: Option<u16>,
+    pub generation: Option<i16>,
+    pub version_group_id: Option<i32>,
     pub release_date: Option<String>,
-    pub platform_id: Option<u16>,
+    pub platform_id: Option<i32>,
     pub is_main_series: Option<bool>,
 }
