+++ pokepi/src/models/type_matchup.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct TypeMatchup {
-    pub attacking_type_id: u8,
-    pub defending_type_id: u8,
+    pub attacking_type_id: i32,
+    pub defending_type_id: i32,
     pub damage_factor: f32,
 }
